requirements.txt
django
gunicorn
django-heroku
pyyaml
uritemplate
djangorestframework
djangorestframework-simplejwt
drf_spectacular
coreapi 
pyyaml